% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABD_test_data.R
\docType{data}
\name{ABD_test_data1}
\alias{ABD_test_data1}
\title{AgeBandDecomposition test data}
\format{
A tibble with 85 rows and 16 variables, as derived from the Excel file.
}
\usage{
ABD_test_data1
}
\description{
A tiny test dataset including example,
the basic reference for AgeBandDecomposition package.
}
\keyword{datasets}
