% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABD_test_data.R, R/stdTRW_test_data.R
\docType{data}
\name{inTRW}
\alias{inTRW}
\title{AgeBandDecomposition test data}
\format{
A tibble with 85 rows and 16 variables, as derived from the Excel file.

A list of two tibbles
}
\usage{
inTRW

inTRW
}
\description{
A tiny test dataset including example,
the basic reference for AgeBandDecomposition package.

A list of two test datasets derived from TRW_readExcel function. The first list-object
is a \code{tibble} that can be used as the argument of \code{stdTRW} function.
}
\keyword{datasets}
