% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/po_test_data.R
\docType{data}
\name{po_test_data}
\alias{po_test_data}
\title{AgeBandDecomposition: pith-offset test data}
\format{
A tibble
}
\usage{
po_test_data
}
\description{
A tiny test dataset showing the structure of a pith-offset (text file) file imported in R
}
\keyword{datasets}
