% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwl_test_data.R
\docType{data}
\name{rwl_test_data}
\alias{rwl_test_data}
\title{AgeBandDecomposition: rwl test data}
\format{
A tibble
}
\usage{
rwl_test_data
}
\description{
A tiny test dataset showing the structure of a rwl file imported in R
}
\keyword{datasets}
