% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRW_readExcel.R
\name{TRW_readExcel}
\alias{TRW_readExcel}
\title{TRW_readExcel (Deprecated)}
\usage{
TRW_readExcel(path, sheet, ageBands, limitFirst20y = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{sheet}{Sheet to read.}

\item{ageBands}{character. Age band window setting.}

\item{limitFirst20y}{logical. Remove first 20 years from each tree.}

\item{verbose}{logical. Print additional information.}
}
\value{
A list of two objects (see original documentation).
}
\description{
\strong{This function is deprecated and will be removed in a future version.}
Please use \code{\link{import_rwl}} instead after converting your data to RWL format.
}
\seealso{
\code{\link{import_rwl}}
}
