% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stdTRW_test_data.R
\docType{data}
\name{inTRW}
\alias{inTRW}
\title{stdTRW test data}
\format{
A list of two tibbles
}
\usage{
inTRW
}
\description{
A list of two test datasets derived from TRW_readExcel function. The first list-object
is a \code{tibble} that can be used as the argument of \code{stdTRW} function.
}
\keyword{datasets}
